ct_dashboard_api: extract customer info logic to api layer
