+++ backend/ct_dashboard_api/src/routes/v1/create_api_key.ts
@@ -1,16 +1,15 @@
+import { ErrorCodeMap } from "@oko-wallet/oko-api-error-codes";
 import { registry } from "@oko-wallet/oko-api-openapi";
 import { ErrorResponseSchema } from "@oko-wallet/oko-api-openapi/common";
 import {
   CreateApiKeySuccessResponseSchema,
   CustomerAuthHeaderSchema,
 } from "@oko-wallet/oko-api-openapi/ct_dashboard";
-import { insertAPIKey } from "@oko-wallet/oko-pg-interface/api_keys";
-import { getCustomerByUserId } from "@oko-wallet/oko-pg-interface/customers";
 import type { OkoApiResponse } from "@oko-wallet/oko-types/api_response";
 import type { APIKey } from "@oko-wallet/oko-types/ct_dashboard";
-import { randomBytes } from "crypto";
 import type { Response } from "express";
 
+import { createApiKeyRequest } from "@oko-wallet-ctd-api/api/customer_info";
 import type { CustomerAuthenticatedRequest } from "@oko-wallet-ctd-api/middleware/auth";
 
 registry.registerPath({
@@ -63,55 +62,14 @@ export async function createApiKey(
   req: CustomerAuthenticatedRequest,
   res: Response<OkoApiResponse<APIKey>>,
 ) {
-  try {
-    const state = req.app.locals;
+  const state = req.app.locals;
 
-    const customerRes = await getCustomerByUserId(state.db, res.locals.user_id);
+  const result = await createApiKeyRequest(state.db, res.locals.user_id);
 
-    if (!customerRes.success) {
-      res.status(500).json({
-        success: false,
-        code: "UNKNOWN_ERROR",
-        msg: customerRes.err,
-      });
-      return;
-    }
-
-    if (customerRes.data === null) {
-      res.status(404).json({
-        success: false,
-        code: "CUSTOMER_NOT_FOUND",
-        msg: "Customer not found",
-      });
-      return;
-    }
-
-    const apiKey = randomBytes(32).toString("hex");
-    const insertRes = await insertAPIKey(
-      state.db,
-      customerRes.data.customer_id,
-      apiKey,
-    );
-
-    if (!insertRes.success) {
-      res.status(500).json({
-        success: false,
-        code: "UNKNOWN_ERROR",
-        msg: insertRes.err,
-      });
-      return;
-    }
-
-    res.status(200).json({
-      success: true,
-      data: insertRes.data,
-    });
-  } catch (error) {
-    console.error("Create API key error:", error);
-    res.status(500).json({
-      success: false,
-      code: "UNKNOWN_ERROR",
-      msg: "Internal server error",
-    });
+  if (!result.success) {
+    res.status(ErrorCodeMap[result.code] ?? 500).json(result);
+    return;
   }
+
+  res.status(200).json(result);
 }
